requirements.txt
git+https://github.com/Pycord-Development/pycord.git#egg=py-cord[voice]
python-Levenshtein>=0.12.0
youtube-dl